요상한 종속성 문제들 해결
