+++ src/containers/index.ts
@@ -4,7 +4,7 @@ import InfoStep from "./InfoStep";
 import QuizStep from "./QuizStep";
 import QuizContainer from "./QuizContainer";
 import FieldVoteStep from "./FieldVoteStep";
-import PopVoteStep from "./FieldVoteStep";
+import PopVoteStep from "./PopVoteStep";
 
 export {
   MainDrawer,
